Don't create deleteTimeout for notification if 'deleteAfter' prop isn't provided or equals zero

+++ src/components/item.jsx
@@ -11,7 +11,9 @@ export default class Notification extends Component {
       deleteAfter,
     } = props;
 
-    this.deleteTimeout = setTimeout(() => deleteNotification(notificationId), deleteAfter);
+    if (deleteAfter) {
+      this.deleteTimeout = setTimeout(() => deleteNotification(notificationId), deleteAfter);
+    }
   }
 
   componentWillUnmount() {
